requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.9
django-cors-headers==4.4.0
djangorestframework==3.15.2
python-decouple==3.8
gunicorn==22.0.0
whitenoise==6.7.0
psycopg2==2.9.11